+++ src/Carousel/index.ts
@@ -1,2 +1,0 @@
-export { SpinoramaCarousel } from './Carousel'
-export * from './Carousel.types'