+++ src/Slideshow/index.ts
@@ -1,2 +1,0 @@
-export { SpinoramaSlideshow } from './Slideshow'
-export * from './Slideshow.types'