+++ src/Wrapper/index.ts
@@ -1,0 +1,2 @@
+export { SpinoramaWrapper } from './Wrapper'
+export * from './Wrapper.types'