+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './Spinorama'
-export * from './Slideshow'
-export * from './Carousel'
+export * from './Wrapper'
 export * from './Item'