+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Both `SpinoramaSlideshow` and `SpinoramaCarousel` components have been merged into a single component called **SpinoramaWrapper**
+
 ## 0.3.8
 - Changed `Grid` to `Box` for the wrapper in the **SpinoramaCarousel** component. This change was made to provide a more flexible and appropriate container for the carousel items